requirements.txt
pyTelegramBotAPI==4.18.0
requests==2.31.0
flask==3.0.3
python-dotenv==1.0.1
